requirements.txt
GPUtil==1.4.0
matplotlib==3.6.0
nltk==3.7
numpy==1.23.3
opencv_python_headless==4.6.0.66
Pillow==9.2.0
requests==2.28.1
scikit_image==0.19.3
skimage==0.0
torch==1.12.1+cu116
torchvision==0.13.1+cu116
